requirements.txt
Flask>=2.0
opencv-python>=4.5
scikit-image>=0.19
imagehash>=4.3
Pillow>=9.0
numpy>=1.21
